+++ src/gromo/containers/growing_dag.py
@@ -66,7 +66,7 @@ class GrowingDAG(nx.DiGraph, GrowingContainer):
 
         self.add_edges_from(edges)
         self.update_nodes(self.nodes, node_attributes)
-        self.update_edges(edges, edge_attributes)
+        self.update_edges(edges, edge_attributes, zero_weights=False)
         self.update_connections(edges)
         self.id_last_node_added = max(len(node_attributes.keys()) - 2, 0)
 
@@ -205,8 +205,22 @@ class GrowingDAG(nx.DiGraph, GrowingContainer):
         """
         return [self.get_node_module(node) for node in nodes]
 
+    def get_all_node_modules(self) -> list[LinearMergeGrowingModule]:
+        """Getter function for all modules attached to nodes
+
+        Returns
+        -------
+        list[LinearMergeGrowingModule]
+            list of modules for all existing nodes
+        """
+        return self.get_node_modules(list(self.nodes))
+
     def add_direct_edge(
-        self, prev_node: str, next_node: str, edge_attributes: dict = {}
+        self,
+        prev_node: str,
+        next_node: str,
+        edge_attributes: dict = {},
+        zero_weights: bool = False,
     ) -> None:
         """Add direct edge to graph, link two nodes with a new module
 
@@ -218,10 +232,14 @@ class GrowingDAG(nx.DiGraph, GrowingContainer):
             outgoing node of edge
         edge_attributes : _type_, optional
             extra attributes of edge, by default {}
+        zero_weights : bool, optional
+            set the weights to zero, by default False
         """
         self.add_edge(prev_node, next_node)
         edges = [(prev_node, next_node)]
-        self.update_edges(edges, edge_attributes=edge_attributes)
+        self.update_edges(
+            edges, edge_attributes=edge_attributes, zero_weights=zero_weights
+        )
         self.update_connections(edges)
 
     def add_node_with_two_edges(
@@ -231,6 +249,7 @@ class GrowingDAG(nx.DiGraph, GrowingContainer):
         next_node: str,
         node_attributes: dict,
         edge_attributes: dict = {},
+        zero_weights: bool = False,
     ) -> None:
         """Add new node to graph, create incoming and outgoing edges with new modules
 
@@ -246,6 +265,8 @@ class GrowingDAG(nx.DiGraph, GrowingContainer):
             attributes of new node
         edge_attributes : dict, optional
             extra attributes of edge, by default {}
+        zero_weights : bool, optional
+            set the weights to zero, by default False
 
         Raises
         ------
@@ -268,7 +289,9 @@ class GrowingDAG(nx.DiGraph, GrowingContainer):
         # TODO: separate functions for different modules, no need to check the type of node
         # self.nodes[new_node].update(node_attributes)
         self.update_nodes([new_node], node_attributes={new_node: node_attributes})
-        self.update_edges(new_edges, edge_attributes=edge_attributes)
+        self.update_edges(
+            new_edges, edge_attributes=edge_attributes, zero_weights=zero_weights
+        )
         self.update_connections(new_edges)
         self.id_last_node_added += 1
 
@@ -347,7 +370,10 @@ class GrowingDAG(nx.DiGraph, GrowingContainer):
                 )
 
     def update_edges(
-        self, edges: list[tuple[str, str]], edge_attributes: dict = {}
+        self,
+        edges: list[tuple[str, str]],
+        edge_attributes: dict = {},
+        zero_weights: bool = False,
     ) -> None:
         """Create new modules for edges based on node types
 
@@ -357,6 +383,8 @@ class GrowingDAG(nx.DiGraph, GrowingContainer):
             list of edges to update modules
         edge_attributes : dict, optional
             extra attributes for edges, by default {}
+        zero_weights : bool, optional
+            set the weights to zero, by default False
         """
         for prev_node, next_node in edges:
             if edge_attributes.get("constant"):
@@ -375,16 +403,21 @@ class GrowingDAG(nx.DiGraph, GrowingContainer):
                 self.nodes[prev_node]["type"] == "linear"
                 and self.nodes[next_node]["type"] == "linear"
             ):
+                new_module = LinearGrowingModule(
+                    in_features=self.nodes[prev_node]["size"],
+                    out_features=self.nodes[next_node]["size"],
+                    use_bias=edge_attributes.get("use_bias", self.use_bias),
+                    device=self.device,
+                    name=f"l{prev_node}_{next_node}",
+                )
+                if zero_weights:
+                    new_module.weight = nn.Parameter(torch.zeros_like(new_module.weight))
+                    if new_module.use_bias:
+                        new_module.bias = nn.Parameter(torch.zeros_like(new_module.bias))
                 self.__set_edge_module(
                     prev_node,
                     next_node,
-                    LinearGrowingModule(
-                        in_features=self.nodes[prev_node]["size"],
-                        out_features=self.nodes[next_node]["size"],
-                        use_bias=edge_attributes.get("use_bias", self.use_bias),
-                        device=self.device,
-                        name=f"l{prev_node}_{next_node}",
-                    ),
+                    new_module,
                 )
                 self[prev_node][next_node]["type"] = "linear"
                 # TODO: set bias to zeros
@@ -504,14 +537,13 @@ class GrowingDAG(nx.DiGraph, GrowingContainer):
             node_module.previous_tensor_m.update()
 
             # Compute optimal possible updates
-            deltas = node_module.compute_optimal_delta(update=True, return_deltas=True)
+            node_module.compute_optimal_delta(update=True, return_deltas=False)
 
             # Compute expressivity bottleneck
             bottleneck[node_module._name] = (
                 node_module.projected_v_goal().clone().detach()
             )  # (batch_size, out_features)
 
-            del deltas
             # TODO: separate to functions that add the hooks and remove them
 
             if constant_module:
@@ -533,18 +565,23 @@ class GrowingDAG(nx.DiGraph, GrowingContainer):
             # node_module.delete_update()
 
         # Reset all hooks
-        for next_node_module in next_node_modules:
-            for parallel_module in next_node_module.previous_modules:
-                parallel_module.reset_computation()
-                # DO NOT delete updates
-                # parallel_module.delete_update(include_previous=False)
+        for node_module in self.get_all_node_modules():
+            if node_module in next_node_modules:
+                for parallel_module in node_module.previous_modules:
+                    parallel_module.reset_computation()
+                    # DO NOT delete updates
+                    # parallel_module.delete_update(include_previous=False)
             # Delete activities
-            next_node_module.delete_update()
+            node_module.delete_update()
 
         if constant_module:
             # Remove constant module if needed
             self.remove_direct_edge(self.root, self.end)
             self.remove_direct_edge(self.root, self.end)
+
+        # TODO: Temporary solution
+        for expansion in actions:
+            expansion.dag = copy.deepcopy(self)
 
         return bottleneck, input_B
 
@@ -834,9 +871,7 @@ class GrowingDAG(nx.DiGraph, GrowingContainer):
         if verbose:
             print("\nExtended Forward DAG...")
         x = self.flatten(x)
-        output: dict[str, tuple[torch.Tensor, torch.Tensor]] = {
-            self.root: (x, torch.empty(x.shape[0], 0))
-        }
+        output: dict[str, tuple[torch.Tensor, torch.Tensor]] = {self.root: (x, None)}
         for node in nx.topological_sort(self):
             if verbose:
                 print(f"{node=}")
@@ -846,11 +881,11 @@ class GrowingDAG(nx.DiGraph, GrowingContainer):
                     print("\t-->", module.name, module)
                 module_input = output[previous_node]
                 activity, activity_ext = module.extended_forward(*module_input)
-                activity_ext = (
-                    activity_ext
-                    if activity_ext is not None
-                    else torch.empty(x.shape[0], module.out_features, device=self.device)
-                )
+                # activity_ext = (
+                #     activity_ext
+                #     if activity_ext is not None
+                #     else torch.empty(0, x.shape[0], module.out_features, device=self.device)
+                # )
 
                 assert activity.shape[1] == self.nodes[node]["size"]
 
@@ -944,6 +979,53 @@ class GrowingDAG(nx.DiGraph, GrowingContainer):
         """
         with torch.no_grad():
             pred = self(x)
+            loss = loss_fn(pred, y)
+
+        if self.out_features > 1:
+            final_pred = pred.argmax(axis=1)
+            correct = (final_pred == y).int().sum()
+            accuracy = (correct / pred.shape[0]).item()
+        else:
+            accuracy = -1
+
+        if with_f1score:
+            if self.out_features > 1:
+                f1score = f1_micro(y.cpu(), final_pred.cpu())
+            else:
+                f1score = -1
+            return accuracy, loss.item(), f1score
+
+        return accuracy, loss.item()
+
+    def evaluate_extended(
+        self,
+        x: torch.Tensor,
+        y: torch.Tensor,
+        loss_fn: Callable,
+        with_f1score: bool = False,
+    ) -> tuple[float, float] | tuple[float, float, float]:
+        """Evaluate extended network on batch
+
+        Important: Assumes that the batch is already on the correct device
+
+        Parameters
+        ----------
+        x : torch.Tensor
+            input features tensor
+        y : torch.Tensor
+            true labels tensor
+        loss_fn : Callable
+            loss function for bottleneck calculation
+        with_f1score : bool, optional
+            calculate f1-score, by default False
+
+        Returns
+        -------
+        tuple[float, float] | tuple[float, float, float]
+            accuracy and loss, optionally f1-score
+        """
+        with torch.no_grad():
+            pred = self.extended_forward(x)
             loss = loss_fn(pred, y)
 
         if self.out_features > 1:
@@ -1086,9 +1168,9 @@ class Expansion:
     def expand(self) -> None:
         """Create new edge or node on the enclosed GrowingDAG"""
         if self.type == "new edge":
-            self.dag.add_direct_edge(self.previous_node, self.next_node, self.edge_attributes)  # type: ignore
+            self.dag.add_direct_edge(self.previous_node, self.next_node, self.edge_attributes, zero_weights=True)  # type: ignore
         elif self.type == "new node":
-            self.dag.add_node_with_two_edges(self.previous_node, self.expanding_node, self.next_node, self.node_attributes, self.edge_attributes)  # type: ignore
+            self.dag.add_node_with_two_edges(self.previous_node, self.expanding_node, self.next_node, self.node_attributes, self.edge_attributes, zero_weights=True)  # type: ignore
 
     def update_growth_history(
         self,
@@ -1135,3 +1217,12 @@ class Expansion:
             del self.dag
             del self.growth_history
             del self.metrics
+
+    def __repr__(self) -> str:
+        if self.type == "new edge":
+            return f"[Expansion]: New edge from {self.previous_node} to {self.next_node}"
+        elif self.type == "new node":
+            return f"[Expansion]: New node {self.expanding_node} from {self.previous_node} to {self.next_node}"
+        elif self.type == "expanded node":
+            return f"[Expansion]: Expanding node {self.expanding_node}"
+        return "[Expansion]: NotImplemented"